#440 extract new classes in their own file as per ARC request
